refacor: card
